+++ sdk/healthinsights/Azure.Health.Insights.RadiologyInsights/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.0.0 (Unreleased)
 
+- Azure HealthInsights client library for .Net. This package contains Microsoft Azure HealthInsights client library.
+
 ### Features Added
+- Initial preview of the Azure.Health.Insights.RadiologyInsights client library.
 
 ### Breaking Changes